+++ benchmarks/express/simple.js
@@ -1,0 +1,9 @@
+
+require.paths.unshift('lib')
+require('express')
+
+get('/', function(){
+  return 'Hello World'
+})
+
+run()