+++ benchmarks/node/static.js
@@ -1,0 +1,14 @@
+
+var fs = require('fs'),
+    http = require('http')
+
+http.createServer(function(req, res) {
+  res.writeHead(200, { 'Content-Type': 'text/plain', 'Transfer-Encoding': 'chunked' })
+  fs.createReadStream('benchmarks/shared/jquery.js')
+    .addListener('data', function(data){
+      res.write(data, 'binary')
+    })
+    .addListener('end', function(){
+      res.end()
+    })
+}).listen(3000, 'localhost')