+++ spec/spec.plugins.cookie.js
@@ -17,13 +17,13 @@ describe 'Express'
         compileCookie('foo', 'bar', options).should.eql 'foo=bar; path=/; domain=.vision-media.ca'
       end
       
-      it 'should currectly format any Date objects'
+      it 'should correctly format any Date objects'
         var options = {
-          expires: new Date('May 25, 1987 11:13:00'),
+          expires: new Date(Date.parse('May 25, 1987 11:13:00 PDT')),
           path: '/foo',
           domain: '.vision-media.ca'
         }
-        compileCookie('foo', 'bar', options).should.eql 'foo=bar; expires=Mon, 25-May-1987 11:13:00 GMT; path=/foo; domain=.vision-media.ca'
+        compileCookie('foo', 'bar', options).should.eql 'foo=bar; expires=Mon, 25 May 1987 18:13:00 GMT; path=/foo; domain=.vision-media.ca'
       end
       
       it 'should convert true to a key without a value'
@@ -55,6 +55,21 @@ describe 'Express'
         var attrs = ' SID     =  1232431234234 ;  data =  foo'
         parseCookie(attrs).should.eql { SID: '1232431234234', data: 'foo' }
       end
+
+      it 'should support complex quoted values'
+        var attrs = 'SID="123456789"; fbs_0011223355="uid=0987654321&name=Test+User"'
+        parseCookie(attrs).should.eql { SID: '123456789', fbs_0011223355: 'uid=0987654321&name=Test User' }
+      end
+      
+      it 'should not override when a duplicate key is found'
+        var attrs = 'SID=1234; SID=9999'
+        parseCookie(attrs).should.eql { SID: '1234' }
+      end
+      
+      it 'should support malformed cookies'
+        var attrs = 'SID'
+        parseCookie(attrs).should.eql {}
+      end
     end
     
     describe 'on'
@@ -73,16 +88,16 @@ describe 'Express'
             this.cookie('SID', '732423sdfs73243', { path: '/', secure: true })
             return ''  
           })
-          get('/user').headers['set-cookie'].should.eql 'SID=732423sdfs73243; path=/; secure'
+          get('/user').headers['Set-Cookie'].should.eql 'SID=732423sdfs73243; path=/; secure'
         end
         
-        it 'should join multiple cookies'
+        it 'should set multiple cookies'
           get('/user', function(){
             this.cookie('SID', '732423sdfs73243', { path: '/', secure: true })
             this.cookie('foo', 'bar')
             return ''  
           })
-          get('/user').headers['set-cookie'].should.eql 'SID=732423sdfs73243; path=/; secure, foo=bar; path=/'
+          get('/user').headers['Set-Cookie'].should.eql 'SID=732423sdfs73243; path=/; secure\r\nSet-Cookie: foo=bar; path=/'
         end
       end
     end